pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "normeval"
version = "0.0.1"
authors = [
  { name="Yannic Bracke", email="yannic.bracke@bbaw.de" },
]
description = "Evaluation tool for normalized text"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "scikit-learn >= 1.1.1",
    "numpy >= 1.22.3",
    "python-Levenshtein >= 0.12.2",
]


[project.urls]
"Homepage" = "https://github.com/ybracke/normeval"